# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BarbrGo - Barber Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /barbers:
    get:
      operationId: getBarbers
      tags: [barbers]
      summary: List barbers with filters
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: area
          in: query
          schema:
            type: string
        - name: minRating
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minExperience
          in: query
          schema:
            type: integer
        - name: homeService
          in: query
          schema:
            type: boolean
        - name: date
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [rating, price, experience, distance]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of barbers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarberListResponse"
    post:
      operationId: createBarber
      tags: [barbers]
      summary: Create barber profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBarberRequest"
      responses:
        "201":
          description: Barber created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
  /barbers/top-rated:
    get:
      operationId: getTopRatedBarbers
      tags: [barbers]
      summary: Get top-rated barbers
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top rated barbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Barber"
  /barbers/{id}:
    get:
      operationId: getBarber
      tags: [barbers]
      summary: Get barber by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barber profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarberDetail"
    put:
      operationId: updateBarber
      tags: [barbers]
      summary: Update barber profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBarberRequest"
      responses:
        "200":
          description: Updated barber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
  /barbers/{id}/availability:
    get:
      operationId: getBarberAvailability
      tags: [barbers]
      summary: Get barber availability for a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: serviceIds
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
  /barbers/{id}/bookings:
    get:
      operationId: getBarberBookings
      tags: [barbers]
      summary: Get barber bookings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Barber bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /barbers/{id}/reviews:
    get:
      operationId: getBarberReviews
      tags: [barbers]
      summary: Get barber reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barber reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /barbers/{id}/services:
    get:
      operationId: getBarberServices
      tags: [barbers]
      summary: Get barber services
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barber services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /barbers/{id}/dashboard:
    get:
      operationId: getBarberDashboard
      tags: [barbers]
      summary: Get barber dashboard summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarberDashboard"
  /barbers/{id}/earnings:
    get:
      operationId: getBarberEarnings
      tags: [barbers]
      summary: Get barber earnings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
      responses:
        "200":
          description: Earnings data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsResponse"

  # Shops
...
  /barbers/{barberId}/services:
    post:
      operationId: addBarberService
      tags: [services]
      summary: Add service to barber
      parameters:
        - name: barberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
components:
  schemas:
...
    CreateBarberRequest:
      type: object
      properties:
        userId:
          type: integer
        shopId:
          type: integer
        experience:
          type: integer
        bio:
          type: string
        city:
          type: string
        area:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        homeService:
          type: boolean
        homeServiceCharge:
          type: number
        workingHoursStart:
          type: string
        workingHoursEnd:
          type: string
        skills:
          type: array
          items:
            type: string
        portfolioImages:
          type: array
          items:
            type: string
      required:
        - userId
        - experience
        - city
        - area
    UpdateBarberRequest:
      type: object
      properties:
        experience:
          type: integer
        bio:
          type: string
        city:
          type: string
        area:
          type: string
        address:
          type: string
        homeService:
          type: boolean
        homeServiceCharge:
          type: number
        workingHoursStart:
          type: string
        workingHoursEnd:
          type: string
        skills:
          type: array
          items:
            type: string
        portfolioImages:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
    Barber:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        avatar:
          type: string
        experience:
          type: integer
        bio:
          type: string
        city:
          type: string
        area:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        homeService:
          type: boolean
        homeServiceCharge:
          type: number
        workingHoursStart:
          type: string
        workingHoursEnd:
          type: string
        skills:
          type: array
          items:
            type: string
        portfolioImages:
          type: array
          items:
            type: string
        rating:
          type: number
        totalReviews:
          type: integer
        totalBookings:
          type: integer
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - experience
        - city
        - area
        - homeService
        - rating
        - totalReviews
        - totalBookings
        - isAvailable
        - createdAt
    BarberDetail:
      allOf:
        - $ref: "#/components/schemas/Barber"
        - type: object
          properties:
            services:
              type: array
              items:
                $ref: "#/components/schemas/Service"
            recentReviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            shopName:
              type: string
    BarberListResponse:
      type: object
      properties:
        barbers:
          type: array
          items:
            $ref: "#/components/schemas/Barber"
        total:
          type: integer
        hasMore:
          type: boolean
      required:
        - barbers
        - total
        - hasMore
    AvailabilityResponse:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
        nextAvailable:
          type: string
      required:
        - date
        - slots
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        available:
          type: boolean
      required:
        - startTime
        - endTime
        - available
    BarberDashboard:
      type: object
      properties:
        totalBookings:
          type: integer
        todayBookings:
          type: integer
        pendingBookings:
          type: integer
        completedBookings:
          type: integer
        totalEarnings:
          type: number
        todayEarnings:
          type: number
        averageRating:
          type: number
        totalReviews:
          type: integer
        recentBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
      required:
        - totalBookings
        - todayBookings
        - pendingBookings
        - completedBookings
        - totalEarnings
        - todayEarnings
        - averageRating
        - totalReviews
        - recentBookings
    EarningsResponse:
      type: object
      properties:
        total:
          type: number
        period:
          type: string
        breakdown:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              amount:
                type: number
            required:
              - label
              - amount
      required:
        - total
        - period
        - breakdown

    # Shops
...
    CreateServiceRequest:
      type: object
      properties:
        serviceTypeId:
          type: integer
        name:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        description:
          type: string
      required:
        - name
        - price
        - durationMinutes
...
    Service:
      type: object
      properties:
        id:
          type: integer
        barberId:
          type: integer
        serviceTypeId:
          type: integer
        name:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        description:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - barberId
        - name
        - price
        - durationMinutes
        - isActive

    # Bookings
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        barberId:
          type: integer
        barberName:
          type: string
        barberAvatar:
          type: string
        userName:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalDuration:
          type: integer
        totalAmount:
          type: number
        discountAmount:
          type: number
        finalAmount:
          type: number
        isHomeService:
          type: boolean
        address:
          type: string
        paymentMethod:
          type: string
          enum: [online, cash]
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        couponCode:
          type: string
        notes:
          type: string
        cancelReason:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - barberId
        - barberName
        - userName
        - services
        - date
        - startTime
        - endTime
        - totalDuration
        - totalAmount
        - discountAmount
        - finalAmount
        - isHomeService
        - paymentMethod
        - paymentStatus
        - status
        - createdAt

    # Reviews
...
    Review:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        barberId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookingId
        - barberId
        - userId
        - userName
        - rating
        - createdAt

    # Offers
...