# lib/api-zod/src/generated/types/barber.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarbrGo - Barber Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Barber {
  id: number;
  userId: number;
  shopId?: number;
  name: string;
  avatar?: string;
  experience: number;
  bio?: string;
  city: string;
  area: string;
  address?: string;
  latitude?: number;
  longitude?: number;
  homeService: boolean;
  homeServiceCharge?: number;
  workingHoursStart?: string;
  workingHoursEnd?: string;
  skills?: string[];
  portfolioImages?: string[];
  rating: number;
  totalReviews: number;
  totalBookings: number;
  isAvailable: boolean;
  createdAt: Date;
}